Make Device::from_fd() public
In some use-cases the process initializing the evdev device doesn't have
the permissions to open the file - instead the fd is handed to it via
e.g. DBus. This is the case for Wayland compositors using logind.

Expose Device::from_fd() as public function so we can initialize a
device purely from the fd.